004: R44 single-precision libm in quantize — 3.78x faster, bit-identical, measured on board
quantize_activations() used fabs()/round(), the double-precision libm
entries, copied literally from upstream to match its promotion behaviour. On
rv32imafc — single-precision FP hardware, no double — that compiled to
__extendsfdf2, round and __fixdfsi per element: three software-emulated
double routines on every activation before every matmul, costing 8.5% of
per-token runtime.

roundf() gives the identical result here, and that was verified rather than
argued: float->double is exact, round() yields an integral value, and the
int8_t conversion matches either way. Swept 2,264,378 values across the
float32 space — zero differences. This is not "improving" the arithmetic,
which runq.c's header forbids; it is the same arithmetic reached more
cheaply.

Measured on hardware:

  quant       8.5%  ->  2.4%        (204.1M -> 54.0M cycles, 3.78x)
  rate        1.14  ->  1.21 tok/s  (+6.1%)
  110 tokens  96.3s ->  90.4s
  accelerable 76.0% -> 81.1%

The arithmetic closes exactly — 150M cycles saved is 6.0 s at 25 MHz and
total runtime fell 5.9 s, with nothing else moving.

Generated text is byte-identical to the previous board run across 110 tokens
and five layers of accumulation, and verify-fw-math still passes bit-exact
against the host reference on all six shapes. Two independent confirmations
that only the cost changed.

This withdraws my earlier recommendation to move activation quantization
into hardware. It was expensive because of soft-float emulation, not because
the operation is costly; what remains is one fdiv.s per element at 2.4%, far
too small to justify RTL. The general lesson is worth keeping: check what
something compiles to before designing silicon to speed it up. A one-word
change recovered 80% of a cost that had already been written up as an
argument for new hardware.

+++ FemtoRV/FIRMWARE/llama2/quantize.c
@@ -2,7 +2,10 @@
  * quantize.c -- implements quantize.h. Verbatim port of upstream
  * llama2.c's runq.c quantize() (also reproduced unchanged in
  * tools/runq_host.c, the golden reference this must match bit-for-bit).
- * See quantize.h for why fabs()/round() (double) are used instead of the
+ * See quantize.h for why fabsf()/roundf() (single) are used -- they are
+ * PROVEN bit-identical to upstream's double-precision fabs()/round() here,
+ * and avoid three software-emulated double routines per element on a CPU
+ * that has no double-precision hardware. The line below originally said
  * float-suffixed forms.
  */
 
@@ -18,7 +21,7 @@ void quantize_activations(int8_t *q, float *s, const float *x, int n, int gs) {
         /* find the max absolute value in the current group */
         float wmax = 0.0f;
         for (int i = 0; i < gs; i++) {
-            float val = fabs(x[group * gs + i]);
+            float val = fabsf(x[group * gs + i]);
             if (val > wmax) {
                 wmax = val;
             }
@@ -31,7 +34,7 @@ void quantize_activations(int8_t *q, float *s, const float *x, int n, int gs) {
         /* calculate and write the quantized values */
         for (int i = 0; i < gs; i++) {
             float quant_value = x[group * gs + i] / scale; /* scale */
-            int8_t quantized = (int8_t) round(quant_value); /* round and clamp */
+            int8_t quantized = (int8_t) roundf(quant_value); /* round and clamp */
             q[group * gs + i] = quantized;
         }
     }
